+++ CHANGELOG.md
@@ -1,14 +1,15 @@
 # CHANGELOG
 
-## v0.0.2-rc.14
+## v0.0.2-rc.15
 
 - Broke out backend to simplify package
 - Updated auth token and refresh handling
 - Trialing persisted cache
-- changed auth redirect to work with incognito
+- Changed auth redirect to work with incognito
 - Updated product image lazy loading
 - Updated auth redirects to return to original page
 - Updated page titles
+- Updated fetch policy
 - Refined Article view rendering & props mapped
 - Fixed issue where pager wasn't calculating maxPage correctly
 - Fixed issue where install prompt was not displaying